Add debug highlight setting
